Mark Goodrich resolved PIMS-740 (Confirm that "sendShipment" works when *destination* is a non-warehouse, and "receiveShipment" works when *origin* is a non-warehouse) saying: This was a bug because creation of a ship or receive transaction was failing (silently) if the source or destination was not a warehouse. For now, I changed the code so that if the source of destination is not a warehouse, the source or destination of the underlying transaction null.
Mark Goodrich resolved PIMS-637 (Confirm that "sendShipment" functionality works correctly when the Origin is a location without an inventory) saying:
Changed this so that if a shipment originates at a non-warehouse, you can "ship" it when logged into the destination warehouse. Also, changed this so that if the destination is a non-warehouse, you can "receive" it when logged into the source warehouse.

I also changed the "receive shipment" page so that if the destination is a non-warehouse, you don't get a list of inventory to accept. Also changed the "send shipment" page so that if the source is a non-warehouse, it doesn't display the verbage about debiting stock from the corresponding warehouse.

git-svn-id: https://dev.pih-emr.org/svn/repository/projects/warehouse/trunk@9874 f0ac6d2b-dfc5-4988-9158-fc266d298d87