Mark Goodrich closed PIMS-739 (Rework Events -- remove "Created" event type) saying:
I removed the "CREATED" event type from the system, and did a major rework of how status is handled... hopefully for the better... 

"Status" is no longer tied to events like it once was. I created a ShipmentStatus object and a ShipmentStatusCode enum (with values Pending, Shipped, and Received).

Shipment now has a getStatus() method, replacing the getMostRecentStatus() method, because "mostRecent" is now somewhat redundant. getStatus no longer need rely on "events" to determine status. As of now, the EventCodes and the ShipmentStatusCodes are similar, but we can diverge from this if we need to.

As of now, everything should operate as it has before (with the exception that no "created" event will be listed).


git-svn-id: https://dev.pih-emr.org/svn/repository/projects/warehouse/trunk@9877 f0ac6d2b-dfc5-4988-9158-fc266d298d87