OBS-1253 Fix regression in updateProductAvailability and updateInventorySnapshots.
As part of restoring test coverage, I added guards to
- ProductAvailabilityService.updateProductAvailability()
- InventorySnapshotService.updateInventorySnapshots()
to prevent them from firing on objects that had no ID.
However, the change also made it impossible for these
methods to clear Product.productCode, which is an
optional field. I'm not sure if this change is causing
the performance problems, but I believe it is a bug.